pyproject.toml
[tool.poetry]
name = "cspi-qe-dev"
version = "0.1.0"
description = ""
authors = ["Michael Pruitt <mpruitt@redhat.com>"]
readme = "README.md"

[tool.poetry.scripts]
ref-sandbox = "cspi_qe_dev.ref_sandbox.api:ref_sandbox"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
click = "^8.1.7"
pytest = "^7.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = [
    "cspi_qe_dev"
]
norecursedirs = [
    "resources"
]

